renamed folder to kodi
